Your post画面作成 (#37)
* branch initial commit

* tailwindcssを導入、header・SignUpを作成

* headerをLinkに修正

* Loginページを作成

* リセットCSSオフ

* パスワード入力欄type変更

* label,inputの属性を修正

* ログイン画面、サインアップ画面を作成

* 新規フォルダ作成_7/15_22:53

* tailwindcss自動並び替え指定_7/16_2:56

* button変更_7/15_2:57

* login/button変更_7/16_3:00

* singup/button変更_7/16_3:03

* input共通化_7/16_3:17

* review_Naoya/pr-11　マージ

* Header追加_7/18_0:18

* ログイン前画面仮置き_7/18/0:28

* 各ページ仮置き_7/18_0:30

* 仮置き_7/18_1:26

* 記事投稿画面作成中_7/18_2:11

* 記事ホーム画面編集_7/18_2:11

* Buttonコンポ、lgサイズ変更_7/18_2:12

* 長方形ボタンコンポ化_7/18_3:10

* 矢印アイコン追加_7/18_3:10

* ホーム仮置き_7/18_3:15

* ボタン修正_7/18_3:16

* 記事投稿画面仮UI完成_7/18_3:17

* 変更

* コンフリクト解消_20250730_01-06

* header削除→layoutに統一_20250805_21-52

* 全体確認　修正_20250805_22-36

* header削除→layoutでかんり_20250805_22-49

* 編集_20250806_17-09

* your-page追加_20250807_11-41

* your-post編集_20250807_11-56

* タイトルcss変更_20250807_12-15

* ヘッダー削除_20250809_10-42

---------
